--- a/.github/workflows/nodejs.yml
+++ b/.github/workflows/nodejs.yml
@@ -21,8 +21,6 @@ jobs:
     steps:
     - name: Checkout
       uses: actions/checkout@v3
-    - name: Install Canvas Dependencies
-      run: sudo apt-get -o Acquire::Retries=3 install build-essential libcairo2-dev libgif-dev libjpeg-dev libpango1.0-dev librsvg2-dev xvfb
     - name: Use Node.js
       uses: actions/setup-node@v3
       with:
@@ -38,11 +36,6 @@ jobs:
       env:
         NODE_ENV: production
 
-    - name: Run Unit Tests for @pixi/node
-      run: xvfb-run --auto-servernum npm run coverage:node
-      env:
-        NODE_ENV: production
-
     - name: Extract Branch Name
       id: branch_name
       if: github.event_name == 'push'
@@ -50,9 +43,6 @@ jobs:
 
     - name: Build for Distribution
       run: npm run dist
-
-    - name: Validate Build for Distribution
-      run: xvfb-run --auto-servernum npm run dist-validate
 
     # Append assets to releases
     - name: Upload Assets to Release