--- a/bundles/pixi.js-node/src/adapter/index.ts
+++ b/bundles/pixi.js-node/src/adapter/index.ts
@@ -1,8 +1,0 @@
-import './polyfills';
-
-export * from './adapter';
-export * from './loadNodeBase64';
-export * from './loadNodeFont';
-export * from './loadNodeTexture';
-export * from './NodeCanvasElement';
-export * from './NodeCanvasResource';