--- a/bundles/pixi.js-node/src/adapter/polyfills.ts
+++ b/bundles/pixi.js-node/src/adapter/polyfills.ts
@@ -1,9 +1,0 @@
-globalThis.requestAnimationFrame = function requestAnimationFrame(fn)
-{
-    return setTimeout(fn, 1000 / 60);
-};
-
-globalThis.cancelAnimationFrame = function cancelAnimationFrame(fn)
-{
-    return clearTimeout(fn);
-};