--- a/bundles/pixi.js-node/test/import.mjs
+++ b/bundles/pixi.js-node/test/import.mjs
@@ -1,5 +1,0 @@
-import { Renderer, Container } from '@pixi/node';
-
-const stage = new Container();
-const renderer = new Renderer();
-renderer.render(stage);