--- a/bundles/pixi.js-node/test/require.cjs
+++ b/bundles/pixi.js-node/test/require.cjs
@@ -1,5 +1,0 @@
-const { Renderer, Container } = require('@pixi/node');
-
-const stage = new Container();
-const renderer = new Renderer();
-renderer.render(stage);