@bigtimebuddy 2023-02-28
PR Build times:
* 5m 14s
* 5m 31s
* 6m 40s
* 5m 56s
* 7m 51s

**Average: 6m 14s**

Previous Build times:
* 7m 9s
* 8m 17s
* 7m 53s
* 6m 0s
* 6m 42s

**Average: 7m 12s**


So it's ~1 minute faster on GitHub Actions.